backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
sqlalchemy
pymysql
cryptography
google-genai
